\name{fastDistSqrd-package}
\alias{fastDistSqrd-package}
\alias{fastDistSqrd}
\docType{package}
\title{
\packageTitle{fastDistSqrd}
}
\description{
\packageDescription{fastDistSqrd}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fastDistSqrd}
\packageIndices{fastDistSqrd}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{fastDistSqrd}

Maintainer: \packageMaintainer{fastDistSqrd}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
